+++ src/Catesta/Resources/AWS/install_modules.ps1
@@ -175,7 +175,7 @@ else {
         catch {
             $message = 'Failed to install {0}' -f $module.ModuleName
             "  - $message"
-            throw $_
+            throw
         }
     }
 }