+++ src/Catesta/Resources/Module/src/Tests/Unit/ExportedFunctions.Tests.ps1
@@ -46,8 +46,8 @@ Describe -Name $ModuleName -Fixture {
                     $help.description.Text | Should -Not -BeNullOrEmpty
                 }
 
-                It -Name 'Includes at least one example' -Test {
-                    $help.examples.example.Count | Should -BeGreaterOrEqual 1
+                It -Name 'Includes an Example' -Test {
+                    $help.examples.example | Should -Not -BeNullOrEmpty
                 }
             }
         }